+++ packages/wotex-binding-http/.formatter.exs
@@ -1,4 +1,5 @@
 [
   inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"],
-  line_length: 100
+  line_length: 100,
+  plugins: [DoctestFormatter]
 ]