+++ packages/wotex-binding-http/test/support/alternate_client.ex
@@ -3,12 +3,15 @@ defmodule Wotex.Binding.HTTP.Test.AlternateClient do
 
   @behaviour Wotex.Binding.HTTP.Client
 
-  @impl true
-  def request(_request, _credential, _config), do: {:error, :unused}
+  @impl Wotex.Binding.HTTP.Client
+  @spec request(term(), term(), term()) :: {:error, :unused}
+  def request(_, _, _), do: {:error, :unused}
 
-  @impl true
-  def subscribe(_request, _credential, _handler, _config), do: {:error, :unused}
+  @impl Wotex.Binding.HTTP.Client
+  @spec subscribe(term(), term(), function(), term()) :: {:error, :unused}
+  def subscribe(_, _, _, _), do: {:error, :unused}
 
-  @impl true
-  def close(_handle, _config), do: :ok
+  @impl Wotex.Binding.HTTP.Client
+  @spec close(term(), term()) :: :ok
+  def close(_, _), do: :ok
 end