+++ packages/wotex-binding-http/test/support/fake_client.ex
@@ -3,19 +3,22 @@ defmodule Wotex.Binding.HTTP.Test.FakeClient do
 
   @behaviour Wotex.Binding.HTTP.Client
 
-  @impl true
+  @impl Wotex.Binding.HTTP.Client
+  @spec request(term(), term(), map()) :: term()
   def request(request, credential, config) do
     send(config.owner, {:client_request, request, credential})
     return(config, :request_return, {:error, :not_configured})
   end
 
-  @impl true
+  @impl Wotex.Binding.HTTP.Client
+  @spec subscribe(term(), term(), function(), map()) :: term()
   def subscribe(request, credential, handler, config) do
     send(config.owner, {:client_subscribe, request, credential, handler})
     return(config, :subscribe_return, {:error, :not_configured})
   end
 
-  @impl true
+  @impl Wotex.Binding.HTTP.Client
+  @spec close(term(), map()) :: term()
   def close(handle, config) do
     send(config.owner, {:client_close, handle})
     return(config, :close_return, :ok)